Feature/move market data endpoint to asset profiles
* Move admin market data listing to asset profiles endpoint

* Add asset profiles response types
