+++ apps/api/src/app/admin/admin.module.ts
@@ -1,6 +1,5 @@
 import { ActivitiesModule } from '@ghostfolio/api/app/activities/activities.module';
 import { TransformDataSourceInRequestModule } from '@ghostfolio/api/interceptors/transform-data-source-in-request/transform-data-source-in-request.module';
-import { ApiModule } from '@ghostfolio/api/services/api/api.module';
 import { BenchmarkModule } from '@ghostfolio/api/services/benchmark/benchmark.module';
 import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
 import { DataProviderModule } from '@ghostfolio/api/services/data-provider/data-provider.module';
@@ -21,7 +20,6 @@ import { QueueModule } from './queue/queue.module';
 @Module({
   imports: [
     ActivitiesModule,
-    ApiModule,
     BenchmarkModule,
     ConfigurationModule,
     DataGatheringQueueModule,
