+++ apps/api/src/app/endpoints/asset-profiles/asset-profiles.module.ts
@@ -1,3 +1,8 @@
+import { ActivitiesModule } from '@ghostfolio/api/app/activities/activities.module';
+import { ApiModule } from '@ghostfolio/api/services/api/api.module';
+import { BenchmarkModule } from '@ghostfolio/api/services/benchmark/benchmark.module';
+import { ExchangeRateDataModule } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.module';
+import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
 import { SymbolProfileModule } from '@ghostfolio/api/services/symbol-profile/symbol-profile.module';
 
 import { Module } from '@nestjs/common';
@@ -7,7 +12,14 @@ import { AssetProfilesService } from './asset-profiles.service';
 
 @Module({
   controllers: [AssetProfilesController],
-  imports: [SymbolProfileModule],
+  imports: [
+    ActivitiesModule,
+    ApiModule,
+    BenchmarkModule,
+    ExchangeRateDataModule,
+    PrismaModule,
+    SymbolProfileModule
+  ],
   providers: [AssetProfilesService]
 })
 export class AssetProfilesModule {}
