+++ libs/ui/src/lib/assistant/assistant.component.ts
@@ -675,7 +675,7 @@ export class GfAssistantComponent implements OnChanges, OnDestroy, OnInit {
     aSearchTerm: string
   ): Observable<SearchResultItem[]> {
     return this.adminService
-      .fetchAdminMarketData({
+      .fetchAssetProfiles({
         filters: [
           {
             id: aSearchTerm,
